Move encoder handler to choc.c
