+++ keyboards/sofle/choc/choc.c
@@ -72,3 +72,25 @@ led_config_t g_led_config = {
 };
 
 #endif
+
+#ifdef ENCODER_ENABLE
+
+bool encoder_update_user(uint8_t index, bool clockwise) {
+    if (index == 0) {
+        if (clockwise) {
+            tap_code(KC_VOLU);
+        } else {
+            tap_code(KC_VOLD);
+        }
+    } else {
+        if (clockwise) {
+            tap_code(KC_PGUP);
+        } else {
+            tap_code(KC_PGDN);
+        }
+    }
+
+    return false;
+}
+
+#endif
